fix(refresh): force assets to refresh via Lookup
Previously it was entirely possible for `ReadAll` to be repeatedly
invoked without refreshing the assets, since `Lookup` only `.get`
instead of `.refresh()`.

Since there is a refresh threshold now, it is better to do a refresh
if the threshold has exceeded, to prevent issues with asset of similar
name being deleted and re-added, or if the asset is removed all
together, which can be detected at the filesystem level instead.

Also fix when status code is non-200 to return as error instead.

Allow the refresh threshold to be specified as arg, though it defaults
to the current 30 seconds.